refactor: split provisioning.go into errors, store, service files
