+++ internal/provisioning/errors.go
@@ -1,0 +1,8 @@
+package provisioning
+
+import "errors"
+
+var (
+	ErrCodeNotFound    = errors.New("provisioning code not found")
+	ErrCodeAlreadyUsed = errors.New("provisioning code already used")
+)
