+++ internal/provisioning/store.go
@@ -6,12 +6,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"log/slog"
-)
-
-var (
-	ErrCodeNotFound  = errors.New("provisioning code not found")
-	ErrCodeAlreadyUsed = errors.New("provisioning code already used")
 )
 
 // Store manages provisioning codes.
@@ -25,29 +19,6 @@ type Store interface {
 	Activate(ctx context.Context, tx *sql.Tx, deviceID, mqttUsername, mqttPassword string) error
 }
 
-// Service orchestrates device provisioning from the user side.
-type Service struct {
-	store  Store
-	logger *slog.Logger
-}
-
-func NewService(store Store, logger *slog.Logger) *Service {
-	if logger == nil {
-		logger = slog.Default()
-	}
-	return &Service{store: store, logger: logger}
-}
-
-// Provision returns an existing unused provisioning code or creates a new one.
-func (s *Service) Provision(ctx context.Context, userID string) (string, error) {
-	code, err := s.store.GetOrCreateCode(ctx, userID)
-	if err != nil {
-		s.logger.Error("provision: get or create code", "user_id", userID, "error", err)
-	}
-	return code, err
-}
-
-// postgresStore implements Store.
 type postgresStore struct {
 	db *sql.DB
 }
